requirements.txt
PyYAML==3.12
yamale==1.8.0